监控指标不需要太高的精度，把浮点型从从f64调整为f32

+++ src/config/metrics.rs
@@ -10,39 +10,39 @@ impl Handler<MetricsQuery> for ConfigActor {
         let list = vec![
             MetricsItem {
                 metrics_type: MetricsKey::ConfigDataSize,
-                record: MetricsRecord::Gauge(self.cache.len() as f64),
+                record: MetricsRecord::Gauge(self.cache.len() as f32),
             },
             MetricsItem {
                 metrics_type: MetricsKey::ConfigListenerClientSize,
-                record: MetricsRecord::Gauge(self.listener.get_listener_client_size() as f64),
+                record: MetricsRecord::Gauge(self.listener.get_listener_client_size() as f32),
             },
             MetricsItem {
                 metrics_type: MetricsKey::ConfigListenerKeySize,
-                record: MetricsRecord::Gauge(self.listener.get_listener_key_size() as f64),
+                record: MetricsRecord::Gauge(self.listener.get_listener_key_size() as f32),
             },
             MetricsItem {
                 metrics_type: MetricsKey::ConfigSubscriberListenerKeySize,
-                record: MetricsRecord::Gauge(self.subscriber.get_listener_key_size() as f64),
+                record: MetricsRecord::Gauge(self.subscriber.get_listener_key_size() as f32),
             },
             MetricsItem {
                 metrics_type: MetricsKey::ConfigSubscriberListenerValueSize,
-                record: MetricsRecord::Gauge(self.subscriber.get_listener_key_size() as f64),
+                record: MetricsRecord::Gauge(self.subscriber.get_listener_key_size() as f32),
             },
             MetricsItem {
                 metrics_type: MetricsKey::ConfigSubscriberClientSize,
-                record: MetricsRecord::Gauge(self.subscriber.get_client_size() as f64),
+                record: MetricsRecord::Gauge(self.subscriber.get_client_size() as f32),
             },
             MetricsItem {
                 metrics_type: MetricsKey::ConfigSubscriberClientValueSize,
-                record: MetricsRecord::Gauge(self.subscriber.get_client_size() as f64),
+                record: MetricsRecord::Gauge(self.subscriber.get_client_size() as f32),
             },
             MetricsItem {
                 metrics_type: MetricsKey::ConfigIndexTenantSize,
-                record: MetricsRecord::Gauge(self.tenant_index.get_tenant_count() as f64),
+                record: MetricsRecord::Gauge(self.tenant_index.get_tenant_count() as f32),
             },
             MetricsItem {
                 metrics_type: MetricsKey::ConfigIndexConfigSize,
-                record: MetricsRecord::Gauge(self.tenant_index.get_config_count().1 as f64),
+                record: MetricsRecord::Gauge(self.tenant_index.get_config_count().1 as f32),
             },
         ];
         Ok(list)
