+++ src/grpc/metrics.rs
@@ -10,15 +10,15 @@ impl Handler<MetricsQuery> for BiStreamManage {
         let list = vec![
             MetricsItem {
                 metrics_type: MetricsKey::GrpcConnSize,
-                record: MetricsRecord::Gauge(self.conn_cache.len() as f64),
+                record: MetricsRecord::Gauge(self.conn_cache.len() as f32),
             },
             MetricsItem {
                 metrics_type: MetricsKey::GrpcConnActiveTimeoutSetItemSize,
-                record: MetricsRecord::Gauge(self.active_time_set.item_size() as f64),
+                record: MetricsRecord::Gauge(self.active_time_set.item_size() as f32),
             },
             MetricsItem {
                 metrics_type: MetricsKey::GrpcConnResponseTimeoutSetItemSize,
-                record: MetricsRecord::Gauge(self.response_time_set.item_size() as f64),
+                record: MetricsRecord::Gauge(self.response_time_set.item_size() as f32),
             },
         ];
         Ok(list)
